feat: make a package, add node export
for #1
