+++ .gitignore
@@ -1,1 +1,3 @@
 .vscode/
+/node_modules
+/.nyc_output
