+++ index.html
@@ -28,7 +28,7 @@
           <a onclick="fetchText('https://raw.githubusercontent.com/unicode-org/cldr/master/docs/ldml/tr35-collation.md')">tr35-collation</a> |
           <a onclick="fetchText('https://raw.githubusercontent.com/unicode-org/cldr/master/docs/ldml/tr35-info.md')">tr35-info</a> |
           <a onclick="fetchText('https://raw.githubusercontent.com/unicode-org/cldr/master/docs/ldml/tr35-keyboards.md')">tr35-keyboards</a>
-          
+
         </div>
         <div class="from-url">
           <div class="subtitle">Load from URL:</div>
@@ -86,12 +86,13 @@
     </div>
   </div>
   <div>
-    Created by <a target="_blank" href="https://github.com/not-dalia">Dalia</a> | 
+    Created by <a target="_blank" href="https://github.com/not-dalia">Dalia</a> |
     Icons made by <a href="https://www.flaticon.com/authors/good-ware" title="Good Ware">Good Ware</a> from <a href="https://www.flaticon.com/" title="Flaticon">www.flaticon.com</a></div>
 </div>
-  
+
 <!-- partial -->
   <script src="./marked.min.js"></script>
+  <script  src="./generate.js?t=1"></script>
   <script  src="./script.js?t=1"></script>
 
   <script>
